finish app

+++ lib/views/widgets/note_item.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -10,6 +11,23 @@ import '../../models/note_model.dart';
 class NoteItem extends StatelessWidget {
   const NoteItem({super.key, required this.note});
   final NoteModel note;
+
+  Widget adaptiveAction(
+      {required BuildContext context,
+      required VoidCallback onPressed,
+      required Widget child}) {
+    final ThemeData theme = Theme.of(context);
+    switch (theme.platform) {
+      case TargetPlatform.android:
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.linux:
+      case TargetPlatform.windows:
+        return TextButton(onPressed: onPressed, child: child);
+      case TargetPlatform.iOS:
+      case TargetPlatform.macOS:
+        return CupertinoDialogAction(onPressed: onPressed, child: child);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -49,7 +67,42 @@ class NoteItem extends StatelessWidget {
               ),
               trailing: IconButton(
                 onPressed: () {
-                  note.delete();
+                  showAdaptiveDialog<String>(
+                    context: context,
+                    builder: (BuildContext context) => AlertDialog.adaptive(
+                      shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(12.0)),
+                      title: const Align(
+                        child: Text(
+                          'Delete note ',
+                        ),
+                      ),
+                      content: const Text(
+                          'Are you sure you want to delete this note?'),
+                      actions: <Widget>[
+                        adaptiveAction(
+                          context: context,
+                          onPressed: () => Navigator.pop(context, 'Cancel'),
+                          child: const Text(
+                            'Cancel',
+                            style: TextStyle(color: Colors.white),
+                          ),
+                        ),
+                        adaptiveAction(
+                          context: context,
+                          onPressed: () {
+                            note.delete();
+                            BlocProvider.of<NotesCubit>(context).fetchAllData();
+                            Navigator.pop(context);
+                          },
+                          child: Text(
+                            'OK',
+                            style: TextStyle(color: Colors.red[600]),
+                          ),
+                        ),
+                      ],
+                    ),
+                  );
 
                   BlocProvider.of<NotesCubit>(context).fetchAllData();
                 },
